mypy.

+++ src/server/connection.py
@@ -216,7 +216,7 @@ class Peers:
     ) -> List[TimestampedPeerInfo]:
         target_peers = [
             TimestampedPeerInfo(
-                peer.host, uint16(peer.port), 0
+                peer.host, uint16(peer.port), uint64(0)
             )
             for peer in self._peers
             if time.time() - self.time_added[peer.get_hash()] < recent_threshold
